Initialization resets, default Deimos parameters, fixes
Made chemfit reset the entire environment when initialize() is called so that multiple initialize() calls cannot be stuck. This is so that the entire environment is setup in one go to avoid future confusion. Added default arm parameters for Deimos. Removed vacuum to air conversion for FITS files, since we should be calculating them in air wavelengths to begin with. For now, made chemfit use normalized continua from ATLAS models