Added support for virtual parameters, redshift
Added the ability to fit for virtual parameters. A virtual parameter is one that does not have a corresponding axis in the model grid, and its effect is instead added to the spectrum on the fly by a preprocessor function that model grid interfaces can now define. Also allowed models to be loaded with additional data that is made available to the preprocessor. To demonstrate this new feature, implemented support for redshift in Grid7. Updated the demo notebook with redshift fitting, and removed MCMC for now. Note that the new feature breaks photometric priors. I just added a NotImplemented() exception for now, will fix later